Run updated rustfmt v0.8.3, add rustfmt.toml (#17)
Run updated rustfmt v0.8.3, add rustfmt.toml

0.8.3 or some earlier version seems to have brought in some changes to
the default style - with more block indentation instead of the earlier
vertical column. Personally, I prefer this given the shorter length of
lines as well as familiarity from different languages.

In addition, a rustfmt.toml is added in this commit that also makes the
array style block, which doesn't seem to be one of the defaults yet.

The function call style should ideally also be Block, but rustfmt seems to
be doing some weird things with that set, so it's commented out
for now.